Fixed data type of p0 element of beam structure (was long, should have been double!).
git-svn-id: https://svn.aps.anl.gov/AOP/oag/trunk/apps/src/elegant@3010 c1b61890-85f3-ea29-e4d6-a272fe79f168

+++ track.h
@@ -324,7 +324,7 @@ typedef struct {
     double p0_original;     /* initial central momentum */
     double **particle;      /* current/final coordinates */
     long n_to_track;        /* initial number of particles being tracked.  Often equal to n_original, but not always. */
-    long p0;                /* current/final central momentum */
+    double p0;                /* current/final central momentum */
     double **accepted;      /* coordinates of accepted particles, with loss info on lost particles */
     long n_accepted;        /* final number of particles being tracked. */
     } BEAM;
